* master: (35 commits)
  ChangeLog: update
  TTerminalQueue::ProcessEvent: Fix condition
  ChangeLog: update
  TTerminalQueue::ProcessEvent: Bugfix: TerminalItem was not NULL'ed
  TTerminalItem::ProcessEvent: process FItem after assert
  Fix typo.
  Fix warnings
  Fix minor warnings
  Fix minor warnings
  Fix compilation
  Sync with WinSCP 5.2.0
  Sync with WinSCP 5.2.0
  Sync with WinSCP 5.2.0
  Sync with WinSCP 5.2.0
  Sync with WinSCP 5.2.0
  Sync with WinSCP 5.2.0
  Sync with WinSCP 5.2.0
  Sync with WinSCP 5.2.0
  Sysutils.cpp: Refactoring
  FtpFileSystem.cpp: Fix AV if connection is lost
  ...

Conflicts:
	src/NetBox/FarPlugin.cpp
	src/NetBox/WinSCPFileSystem.cpp
	src/core/SessionInfo.cpp

+++ src/NetBox/FarPlugin.cpp
@@ -1594,7 +1594,7 @@ void TCustomFarPlugin::HandleException(Exception * E, int /*OpMode*/)
 UnicodeString TCustomFarPlugin::GetMsg(intptr_t MsgId)
 {
   TFarEnvGuard Guard;
-  UnicodeString Result = FStartupInfo.GetMsg(&MainGuid, MsgId);
+  UnicodeString Result = FStartupInfo.GetMsg(&MainGuid, (int)MsgId);
   return Result;
 }
 //---------------------------------------------------------------------------
